+++ src/integTest/groovy/com/moowork/gradle/gulp/GulpRule_integTest.groovy
@@ -11,7 +11,7 @@ class GulpRule_integTest
         given:
         writeBuild( '''
             plugins {
-                id 'com.moowork.gulp'
+                id 'nebula.gulp'
             }
 
             node {