+++ src/test/groovy/com/moowork/gradle/node/NodeExtensionTest.groovy
@@ -8,7 +8,7 @@ class NodeExtensionTest
     def "check default values for extension"()
     {
         when:
-        this.project.apply plugin: 'com.moowork.node'
+        this.project.apply plugin: 'nebula.node'
         def ext = NodeExtension.get( this.project )
 
         then: